requirements.txt
pandas==2.2.3
pyarrow==18.0.0
numpy==1.26.4
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pymongo==4.10.1
python-dotenv==1.0.1
gdown==5.2.0
matplotlib==3.9.3
seaborn==0.13.2
plotly==5.24.1
lxml==5.3.0
jupyter==1.1.1
notebook==7.3.1
ipykernel==6.29.5
python-docx==1.1.2
python-pptx==1.0.2
openpyxl==3.1.5
tabulate==0.9.0
tqdm==4.67.1
docx2pdf==0.1.8
pytest==8.3.4
